// client/src/app/models/museum.ts
export interface Museum {
  museo_id?: number;
  museo_nombre?: string;
  museo_tematica_n1?: string;
  museo_fecha_fundacion?: Date;
  museo_adscripcion?: string;
  estado_id?: number;
  nom_ent?: string;
  municipio_id?: number;
  nom_mun?: string;
  localidad_id?: number;
  nom_loc?: string;
  museo_calle_numero?: string;
  museo_colonia?: string;
  museo_cp?: number;
  museo_telfono1?: string;
  pagina_web?: string;
  pagina_web2?: string;
  email?: string;
  twitter?: string;
  gmaps_longitud?: number;
  gmaps_latitud?: number;
  link_sic?: string;
  fecha_mod?: Date;
  key?: any;

}